BucketInfo should take list of strings, not jsonarray for keys Never return null from operations. For ExceptionHandler case, return 0-status response. Add missing getallowmult() to rawbucketinfo DefaultHttpResponse never null headers; null checks for issuccess/iserror Tests for response package Soem tests for raw package

+++ src/test/java/com/basho/riak/client/itest/ITestRawBasic.java
@@ -124,7 +124,7 @@ public class ITestRawBasic {
         bucketresp = c.listBucket(BUCKET);
         assertTrue(bucketresp.isSuccess() && bucketresp.hasBucketInfo());
         bucketInfo = bucketresp.getBucketInfo();
-        assertEquals(nval + 1, bucketInfo.getNVal());
+        assertEquals(nval + 1, bucketInfo.getNVal().intValue());
         assertEquals(CHASH_MOD + ":" + CHASH_FUN, bucketInfo.getCHashFun());
     }
